+++ tournament/src/presentation/api/classes/HistoryDTO2JSON.ts
