+++ tournament/src/presentation/api/classes/TournamentDTO2JSON.ts
