+++ tournament/src/presentation/api/classes/errors/error.ts
