--- tournament/src/presentation/api/errors/ErrorGenerator.ts
@@ -1,14 +1,0 @@
-// import { ErrorSchema } from "../schemas/ErrorSchema";
-
-
-// export class ErrorGenerator {
-
-// 	static createError(code: string, message: string, statusCode: number, ): ErrorSchema {
-
-// 	}
-// 	static notFound(): ErrorSchema {
-// 		return {
-// 			code: ""
-// 		}
-// 	}
-// }
