+++ tournament/src/presentation/api/routes/route.ts
@@ -1,7 +1,7 @@
 import { FastifyInstance } from "fastify";
-import { TournamentRoutes } from "./tournament/routes/TournamentRoutes";
-import { HistoryRoutes } from "./history/routes/HistoryRoutes";
-import { ParticipantRoutes } from "./participant/routes/ParticipantRoutes";
+import { TournamentRoutes } from "./tournaments/TournamentRoutes";
+import { HistoryRoutes } from "./histories/HistoryRoutes";
+import { ParticipantRoutes } from "./participants/ParticipantRoutes";
 
 export function TournamentAPIRoutes(fastify: FastifyInstance) {
 	fastify.register(TournamentRoutes);
