+++ tournament/src/presentation/api/routes/tournaments/GetTournamentHistoryRoute.ts
@@ -1,13 +1,13 @@
 import { Type } from "@sinclair/typebox";
 import { FastifyInstance } from "fastify";
-import { NotFoundError, NotImplementedError } from "../../errors/error";
+import { NotFoundError, NotImplementedError } from "../../classes/errors/error";
 import { HistorySchema } from "../../schemas/HistorySchema";
 import { TournamentSchema } from "../../schemas/TournamentSchema";
 import { ToStatic } from "../../../types/ToStatic";
 import { TournamentApplicationService } from "../../../../application/service/tournament/TournamentApplicationServiceFacade";
 import { PrismaRepositoryFactory } from "../../../../infrastructure/Prisma/PrismaReopsitoryFactory";
 import { PrismaClientProvider } from "../../../../infrastructure/Prisma/PrismaClientProvider";
-import { HistoryDTO2JSON } from "../HistoryDTO2JSON";
+import { HistoryDTO2JSON } from "../../classes/HistoryDTO2JSON";
 import { DIContainer } from "../../../classes/DIContainer";
 
 const description = `
